pyproject.toml
[project]
name = "learning-agent"
version = "0.1.0"
description = "Add your description here"
readme = "plan/README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.8.3",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.1.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.20",
    "uvicorn>=0.42.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]
